pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neuralcache"
version = "0.1.0"
description = "NeuralCache: narrative + stigmergic reranker for RAG (Context-Use@K uplift)."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Carnot Engine", email = "hello@carnotengine.com"}]
keywords = ["RAG", "reranker", "LLM", "information retrieval", "narrative coherence", "stigmergy"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

dependencies = [
  "fastapi>=0.110,<1.0",
  "uvicorn[standard]>=0.24",
  "numpy>=1.26",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "scipy>=1.11",
  "typer>=0.12",
  "orjson>=3.9",
  "python-dateutil>=2.9",
  "rich>=13.7",
  # optional adapters (safe to import conditionally)
  "langchain-core>=0.2; python_version >= '3.11'",
  "llama-index-core>=0.10; python_version >= '3.11'",
]

[project.optional-dependencies]
test = ["pytest>=8.0", "pytest-cov>=5.0"]
dev = ["ruff>=0.4", "mypy>=1.8", "pre-commit>=3.6", "httpx>=0.27"]
ops = ["prometheus-client>=0.20", "requests>=2.32"]

[project.scripts]
neuralcache = "neuralcache.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
indent-width = 4
target-version = "py311"
fix = true

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "DTZ", "ISC", "PIE", "Q", "RET", "SIM", "PTH"]
ignore = ["E203"]
preview = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
strict_optional = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
